backend/requirements.txt
langchain
langchain_core
langchain_anthropic
langchain_openai
tavily_python
langchain_community
langgraph
pyairbnb
geopy
python-dotenv
fastapi
uvicorn
jinja2
pydantic
sendgrid
tavily_python
langchain_groq
fast-flights
PyPDF2
beautifulsoup4
googlemaps
docling
arxiv
needle-python
openai
reportlab
aiohttp
ipython
primp==0.11.0
python-docx 
docx2pdf
pdfkit
langfuse